chore: fix all ruff lint errors (CI green)
- Auto-fix unused imports across modules and tests
- Replace deprecated datetime.timezone.utc with UTC
- Replace generator with set comprehension; dict comprehension with dict()
- Rename single-letter loop variables (l → label, s → _s)
- Remove unused vol assignment in detect_regimes
- Remove unnecessary int() cast
- Configure ruff to ignore RUF001-003 (Japanese × multiplication sign),
  N806 (uppercase L for leverage convention), and UP042 (StrEnum)

74 tests still passing. ruff check clean.

+++ src/leverage_survival_lab/analysis/metrics.py
@@ -64,5 +64,5 @@ def risk_of_ruin(p_win: float, b: float, *, ruin_pct: float = 0.5, n_units: int 
     if edge <= 0:
         return 1.0
     base = (1 - edge) / (1 + edge)
-    units = n_units if n_units is not None else int(round(1.0 / ruin_pct))
+    units = n_units if n_units is not None else round(1.0 / ruin_pct)
     return float(base ** units)
